a function that returns the length of a string.

+++ 0x05-pointers_arrays_strings/2-strlen.c
@@ -5,7 +5,7 @@
  * @s: The string to get the length of.
  * Return : returns length of @s
 */
-int _strlen(char *s)
+size_t _strlen(char *s)
 {
 size_t length = 0;
 while (*s++)
